requirements.txt
flask
pandas
fitz
numpy
pytesseract
docx
pymupdf
pdf2docx
python-pptx
python-docx